Add method for attaching and detaching label
Add methods for attaching and detaching labels to specified host and
interface.

Add method for searching if label is present on some interface.

Add configuration for collisions. When label is attached to other
interface than candidate is meant for, the collision flag indicates,
if the interface should relabeled to newly discovered configuration.

Add method for checking label candidate and assign it to proper
interface.